Move analyser into registry
