+++ talondoc/sphinx/__init__.py
@@ -6,8 +6,8 @@ from sphinx.application import BuildEnvironment, Sphinx
 from typing_extensions import TypeGuard
 
 from .. import __version__
-from ..analyzer.standalone import analyse_package
 from ..registry import NoActiveFile, NoActivePackage, NoActiveRegistry
+from ..registry.analyzer import analyse_package
 from ..util.logging import getLogger
 from .domains import TalonDomain
 from .typing import TalonPackage
